trace_shed: Drop useless and wrong WEXITSTATUS(status)
The thread return status is either -1 or pointer to the table. So there
is no point in passing it to WEXITSTATUS() and priting the value.

The code started to fail to compile after glibc commit:
https://sourceware.org/git/?p=glibc.git;a=commitdiff;h=b49ab5f4503f36dcbf43f821f817da66b2931fe6;hp=5f5682b9654101ccaf375c2814cbddbe6033a725

trace_sched.c:425:16: error: invalid operands to binary & (have
'thread_sched_t * {aka struct <anonymous> *}' and 'int')
      thrd_ndx, WEXITSTATUS(status));

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>